some changes

+++ util/delay_functions.py
@@ -1,7 +1,7 @@
 import asyncio
 
 
-async def delay(delay_seconds: int) -> int:
+async def delay(delay_seconds: float) -> float:
     print(f'sleeping for {delay_seconds} second(s)')
     await asyncio.sleep(delay_seconds)
     print(f'finished sleeping for {delay_seconds} second(s)')
